[#issue22]: Refactor and restucture server module
