+++ WizardServer/src/main/java/at/aau/ase/androidnetworkwrapper/networking/dto/BaseMessage.java
@@ -1,0 +1,4 @@
+package at.aau.ase.androidnetworkwrapper.networking.dto;
+
+public abstract class BaseMessage {
+}
