+++ WizardServer/src/main/java/at/aau/ase/androidnetworkwrapper/networking/kryonet/KryoNetComponent.java
@@ -1,4 +1,4 @@
-package at.aau.ase.wizard.networking.kryonet;
+package at.aau.ase.androidnetworkwrapper.networking.kryonet;
 
 public interface KryoNetComponent {
