+++ WizardServer/src/main/java/at/aau/ase/androidnetworkwrapper/networking/kryonet/NetworkConstants.java
