--- WizardServer/src/main/java/at/aau/ase/wizard/networking/dto/BaseMessage.java
@@ -1,4 +1,0 @@
-package at.aau.ase.wizard.networking.dto;
-
-public abstract class BaseMessage {
-}
