--- WizardServer/src/main/java/at/aau/ase/wizard/networking/kryonet/NetworkConstants.java
@@ -1,8 +1,0 @@
-package at.aau.ase.wizard.networking.kryonet;
-
-class NetworkConstants {
-
-    public static final int TCP_PORT = 54555;
-    public static final int UDP_PORT = 54777;
-
-}
