--- app/src/main/java/at/aau/ase/wizard/networking/Callback.java
@@ -1,10 +1,0 @@
-package at.aau.ase.wizard.networking;
-
-/**
- * Used for callbacks.
- */
-public interface Callback<T> {
-
-    void callback(T argument);
-
-}
