--- app/src/main/java/at/aau/ase/wizard/networking/dto/TextMessage.java
@@ -1,18 +1,0 @@
-package at.aau.ase.wizard.networking.dto;
-
-public class TextMessage extends BaseMessage {
-
-    public TextMessage() {
-    }
-
-    public TextMessage(String text) {
-        this.text = text;
-    }
-
-    public String text;
-
-    @Override
-    public String toString() {
-        return String.format("TextMessage: %s", text);
-    }
-}
