--- app/src/main/java/at/aau/ase/wizard/networking/kryonet/KryoNetComponent.java
@@ -1,12 +1,0 @@
-package at.aau.ase.wizard.networking.kryonet;
-
-public interface KryoNetComponent {
-
-    /**
-     * Register a class for serialization.
-     *
-     * @param c
-     */
-    void registerClass(Class c);
-
-}
