--- app/src/main/java/at/aau/ase/wizard/networking/kryonet/NetworkClientKryo.java
@@ -1,44 +1,0 @@
-package at.aau.ase.wizard.networking.kryonet;
-
-import at.aau.ase.wizard.networking.Callback;
-import com.esotericsoftware.kryonet.Client;
-import com.esotericsoftware.kryonet.Connection;
-import com.esotericsoftware.kryonet.Listener;
-
-import java.io.IOException;
-
-import at.aau.ase.wizard.networking.dto.BaseMessage;
-import at.aau.ase.wizard.networking.NetworkClient;
-
-public class NetworkClientKryo implements NetworkClient, KryoNetComponent {
-    private Client client;
-    private Callback<BaseMessage> callback;
-
-    public NetworkClientKryo() {
-        client = new Client();
-    }
-
-    public void registerClass(Class c) {
-        client.getKryo().register(c);
-    }
-
-    public void connect(String host) throws IOException {
-        client.start();
-        client.connect(5000, host, NetworkConstants.TCP_PORT, NetworkConstants.UDP_PORT);
-
-        client.addListener(new Listener() {
-            public void received(Connection connection, Object object) {
-                if (callback != null && object instanceof BaseMessage)
-                    callback.callback((BaseMessage) object);
-            }
-        });
-    }
-
-    public void registerCallback(Callback<BaseMessage> callback) {
-        this.callback = callback;
-    }
-
-    public void sendMessage(BaseMessage message) {
-        client.sendTCP(message);
-    }
-}
